+++ connecteur-type/SAEConnecteur.class.php
@@ -39,6 +39,7 @@ abstract class SAEConnecteur extends  Connecteur {
 	
 	abstract public function getErrorString($number);
 
-    abstract public function getLastErrorCode();
-
+    public function getLastErrorCode() {
+        return 0;
+    }
 }